requirements.txt
flask
python-dotenv
aiohttp
requests
langchain==0.2.9
langchain-community==0.2.7
python-dotenv==1.0.0
faiss-cpu
sentence-transformers